+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 - Added customer payment sepa support
 - Added product delivery time manuell and fixed value
+- Fixed customer order shipping and billing additional text handling
 - Changed payment pull error tolerance
 
 1.3.0